moved motion back out of gamelib
