+++ app/gamelib/src/game/mod.rs
@@ -1,5 +1,5 @@
-use ogc_rs::println;
-use ogc_rs::print;
+// use ogc_rs::println;
+// use ogc_rs::print;
 use crate::game_state::{changes::ChangeProvider, GameState};
 use crate::game_states::{GameStateName, GameStateFactory};
