+++ app/gamelib/src/game_state/changes/mod.rs
@@ -1,5 +1,4 @@
 pub mod controls;
-pub mod motion;
 
 use core::time::Duration;
 use controls::Controls;
