+++ app/gamelib/src/game_state/systems/actions.rs
@@ -1,11 +1,11 @@
-use ogc_rs::print;
-use ogc_rs::println;
+// use ogc_rs::print;
+// use ogc_rs::println;
 use crate::game_state::GameState;
 use crate::game_state::components::motion::Velocity;
 use rand::rngs::SmallRng;
 use rand::RngCore;
 use rand::SeedableRng;
-use crate::game_state::changes::motion::Direction;
+use crate::game_state::changes::controls::Direction;
 
 /**
  * Stop the game from running if the home buttons is pressed.
@@ -19,20 +19,17 @@ pub fn system_exit_action(state: &mut GameState) {
 /**
  * Create velocity in random directions if 'one' is pressed.
  */
-pub fn system_shake_action(state: &mut GameState) {
+pub fn system_stop_action(state: &mut GameState) {
     if state.changes.controls.wii_mote_control[0].one_button_down {
-        let mut small_rng = SmallRng::seed_from_u64(10u64);
         for (_id, velocity) in state.world.query_mut::<&mut Velocity>() {
-            // velocity.x += small_rng.next_u32() as f32 / u32::MAX as f32 * 0.2 - 0.1;
             velocity.x = 0.0;
-            // velocity.y += small_rng.next_u32() as f32 / u32::MAX as f32 * 0.2 - 0.1;
             velocity.y = 0.0;
             velocity.z = 0.0;
         }
     }
 }
 
-pub fn system_shake_action_2(state: &mut GameState) {
+pub fn system_shake_action(state: &mut GameState) {
     match &state.changes.controls.wii_mote_control[0].motion {
         None => (),
         Some(motion) => {
