+++ app/gamelib/src/game_state/systems/system_name.rs
@@ -9,8 +9,8 @@ use super::actions::*;
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
 pub enum SystemName {
     ExitAction,
+    StopAction,
     ShakeAction,
-    ShakeAction2,
     IntegrateMotion,
     BounceBounds,
 }
@@ -20,8 +20,8 @@ impl SystemName {
     pub const fn to_function(&self) -> &'static dyn Fn(&mut GameState) {
         match self {
             SystemName::ExitAction => &system_exit_action,
+            SystemName::StopAction => &system_stop_action,
             SystemName::ShakeAction => &system_shake_action,
-            SystemName::ShakeAction2 => &system_shake_action_2,
             SystemName::IntegrateMotion => &system_integrate_motion,
             SystemName::BounceBounds => &system_bounce_bounds,
         }
