+++ app/gamelib/src/game_states/bouncing_cubes.rs
@@ -14,8 +14,8 @@ use rand::SeedableRng;
 pub fn build() -> GameState {
     let mut state = GameState::new();
     state.add_system(SystemName::ExitAction);
+    state.add_system(SystemName::StopAction);
     state.add_system(SystemName::ShakeAction);
-    state.add_system(SystemName::ShakeAction2);
     state.add_system(SystemName::IntegrateMotion);
     state.add_system(SystemName::BounceBounds);
     batch_spawn_entities(&mut state.world, 10);
